+++ src/interface/rest-controller/src/lib.rs
@@ -1,7 +1,7 @@
 pub mod handler;
-pub mod route;
+pub mod router;
 pub mod schema;
 
 // ルーティング設定を公開
-pub use route::health::configure as configure_health;
-pub use route::user::configure as configure_users;
+pub use router::health::configure as configure_health;
+pub use router::user::configure as configure_users;
