+++ src/interface/rest-controller/src/router/health.rs
